requirements.txt
colorama==0.4.6
python-dotenv==1.0.0
requests==2.28.2
beautifulsoup4==4.12.2
google==3.0.0
panel